cloud_reid_worker/requirements.txt
google-cloud-pubsub==1.0.2
nltk==3.4.5
scikit-learn==0.21.3
joblib==0.14.0
scipy==1.3.1